Removes post form from admin panel
This was my actual intent when I first began working on the panel. The problem at the time was authenticating staff members using the old PANEL_PASS value. 

Now that staff authentication is handled via login, the admin form can be moved to the index and replies, allowing admins to post and reply to a thread like a normal user (a feature which makes more sense, streamlines ease of use for staff and has been long desired).