Log - More generation options.
For individual index/thread generation.

Log->generate("index",1) will generate (not print) the first index page
(0).
Log->generate("thread",1) will generate (not print) the first thread.

Thread generation also accepts an extra argument for inIndex. If true,
it will attempt to find it in the cache (generated from Index) and use
that. If not, it will be generated from scratch with inIndex's value.