I invoke @RePod - Allow admins edit boardlist/globalnews from panel
This is a neat feature I thought up today, adds an interface to the
admin panel to allow the editing of global board news and board lists
from the panel. The `valid()` call could eventually be replaced when the
user class is finished to allow editing the list/news to be a permission
granted to certain users.

RePod: Test.php needs to be updated to chmod SOMETHING so that admin.php
or news.php can actually write to the file and I am not familiar with
file permissions, so I am asking for your wisdom.

Future application: This could be reused to allow the editing of a front
page news post of some sort, but we're far away from that. more later.

To-do for completion:
- [ ] Fix that spelling error for the news header
- [ ] test.php chmods whatever needs the permission