Resolve "FILE DELETED" for posts with no image
Old regist didn't store a value for `ext` at all if there was no image
attatchment, so **/_core/thread/image.php** checked to see if `ext`
existed for a post. If it had a value but no file could be found, it
returned "IMAGE DELETED". With new regist, `.` is stored in `ext` even
if  there is no file, which tripped up **/_core/thread/image.php**.

This will, however, mess with posts where images are actually deleted,
but that's not regist's fault, that needs to be resolved in a deletion
redesign.