Switch to PDO
This forces an update from mysql to PDO support, prepared statements
included and all. This should make things a lot safer overall as there's
no longer a need to escape strings, and any SQL queries without
hardcoded input are prepared now.

This is all through our own next to useless wrapper that I made which
definitely will need to be revisited as this is like the switch to
classes all over again for me.